+++ lib/random-rails.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require "random-rails/version"
+require "random-rails/configuration"
 require "random-rails/active_record"
 
 module RandomRails