+++ spec/lib/random/active_record_spec.rb
@@ -8,4 +8,20 @@ RSpec.describe ActiveRecord do
   end
 
   it { expect(Dummy).to respond_to(:random) }
+
+  describe "return types" do
+    subject { Person.random }
+
+    it "returns single object by default" do
+      expect(subject).to be_a(Person).or be_nil
+    end
+
+    it "returns single object when count is 1" do
+      expect(subject).to be_a(Person).or be_nil
+    end
+
+    it "returns relation when count is greater than 1" do
+      expect(Person.random(count: 3)).to be_a(ActiveRecord::Relation)
+    end
+  end
 end